refactoring Warehouse structure
Working on warehouse specification flexibility. To achieve it, implemented WarehouseSpec class with Map filed as specification. That way, warehouses are loosely coupled with specifications and there is a possibility to implement different specification for different warehouses.

Created separated controllers and services for specification to achieve better follow of OOP principles.